Minor improvements to wxlive.PlotCanvas.
The PlotCanvas now provides a popup menu by default, that can be used to set
the grid, legend, scaling of axes (not complete) and allows you to "turn off"
plotting of each curve individually.

The README file was updated to give a bit more information. Still not much
though.

The graph.py example has received a minor update to support the popup menu of
the PlotCanvas. Namely, the View menu was removed, because I have not figured
out how to cleanly make the checked items consistent with the ones in the
popup. Also, the StaticText now displays a bit more information.