Massive overhaul, "compatibility is for losers".
Removed the concept of updaters etc. Everything has changed.

You can define a wxlive.Variable, which runs a function to get its value
(fget) and a function to set its value (fset) and sends events to listeners
when the value is changed.

A wxlive.Variable has a property called value. If its value is set, the fset
function is called. If its value is retrieved, the fget function is called.
This is identical to using the methods set_value and get_value.

A wxlive.Variable has a method called get_time_value_pair, which returns a
tuple where the first item is the number of seconds since the epoch minus the
time_offset of the wxlive.Variable that the value was last updated, and the
second item is the last updated value.

A wxlive.Variable can be constantly updated in its own thread, using the
start method (there is also a stop method, and an is_active method to verify
the state of the thread). If the thread is active, retrieving the value of a
property returns the value since the last update by the thread. In other
words, letting the thread run and update the value each second, and then
retrieve the value ten times in one second, will not update the value ten
times. If the thread is not active, the value is only updated whenever it is
called for (i.e. in that case, the value would be updated ten times).

A wxlive.Variable sends events that have a time and value attribute. The time
attribute is the number of seconds since the epoch, minus the time_offset of
the wxlive.Variable.

Any widget can be a listener, using the wxlive.make_listener function and
providing it with an event handler.

If a widget requires a history of values (such as a plotting widget), it
should provide a history buffer itself. This seems fair, since the need for a
history is limited only to a few widgets and by no means a general thing.

Standard, useful widgets will be added (including strip charts, although wxmpl
seems nice for that) at a later time. Meanwhile, make_listener should do the
trick, mostly.

Example will be commited shortly, I hope.