Added reply_is_new_value = False keyword argument
If reply_is_new_value is True, it is assumed that the set function (given with
the fset keyword argument) returns the actual new value. Thus instead of
updating the wxlive.Variable's value with the value that was assigned, it is
updated with the value returned by the set function.