Changed update behaviour, and added method notify_listeners(skip_listener=None)
Update behaviour used to not update timestamp nor notify listeners
when the value did not change.  This is stupid.  Now it updates the
timestamp if there was a get function defined and used, and will
always notify the listeners.

The method notify_listeners(skip_listener=None) is now exposed.  This
will notify all listeners of the variable's value (without updating),
but will skip the listener widget in skip_listener.  E.g. if a
listener is responsible for triggering the event, it does not need to
receive the change event.  This keyword is mostly used internally by
setter widgets.